requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-ckeditor
Pillow
django-cors-headers

